pyproject.toml

[tool.poetry]
name = "datahub-postgres-lineage"
version = "0.1.0"
description = "Extract table lineage from Postgres views"
authors = ["Contiamo <developers@contiamo.com>"]

[tool.poetry.dependencies]
python = ">=3.8.1,<3.11"
acryl-datahub = {extras = ["sqlalchemy"], version = "^0.9.3.2"}
pydantic = "^1.10.2"
psycopg2-binary = "^2.9.5"
geoalchemy2 = "^0.12.5"


[tool.poetry.dev-dependencies]
black = "^22.12.0"
mypy = "^0.991"
pytest = "^7.2.0"
flake8 = "^6.0.0"
ipython = "^8.7.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.pytest.ini_options]
testpaths = [
    "tests/unit",
]

[tool.mypy]
plugins = [
    "sqlmypy",
    "pydantic.mypy"
]
exclude = "^(venv|build|dist)/"
ignore_missing_imports = true
strict_optional = true
check_untyped_defs = true
disallow_incomplete_defs = true
disallow_untyped_decorators = true
warn_unused_configs = true
disallow_untyped_defs = false